return JsonResult in service for collection related

+++ app/Http/Service/CollectionService.php
@@ -11,7 +11,9 @@ namespace App\Http\Service;
 
 use App\Http\Dao\CollectionDao;
 use App\Http\Dao\SkuDao;
+use App\Http\Enum\StatusCode;
 use App\Http\Model\Collection;
+use App\Http\Util\JsonResult;
 use Illuminate\Support\Facades\Log;
 
 /**
@@ -36,22 +38,24 @@ class CollectionService
      * 创建收藏
      *
      * @param array $req
-     * @return bool
+     * @return JsonResult
      */
     public function createCollection(array $req)
     {
+        if (empty($req["userId"]) || empty($req["skuId"])) return new JsonResult(StatusCode::PARAM_LACKED);
         $collection = new Collection();
         $collection->user_id = $req["userId"];
         $collection->sku_id = $req["skuId"];
         $result = $this->collectionDao->insert($collection);
-        return $result;
+        if ($result) return new JsonResult();
+        return new JsonResult(StatusCode::SERVER_ERROR);
     }
 
     /**
      * 获取用户收藏
      *
      * @param array $req
-     * @return mixed
+     * @return JsonResult
      */
     public function getCollectionList(array $req)
     {
@@ -62,7 +66,7 @@ class CollectionService
             $sku = $this->skuDao->findById($collection->sku_id);
             array_push($result, $sku);
         }
-        return $result;
+        return new JsonResult(StatusCode::SUCCESS, $result);
     }
 
     /**
@@ -70,16 +74,18 @@ class CollectionService
      *
      * @param string $userId
      * @param array $skuIds
-     * @return mixed
+     * @return JsonResult
      */
     public function removeCollection(string $userId, array $skuIds)
     {
+        if (empty($req["userId"]) || empty($req["skuIds"])) return new JsonResult(StatusCode::PARAM_LACKED);
         if (sizeof($skuIds) == 1)
             $result = $this->collectionDao->deleteByUserSkuId($userId, $skuIds[0]);
         else {
             $result = $this->collectionDao->deleteByUserSkuIdsIn($userId, $skuIds);
         }
-        return $result;
+        if ($result) return new JsonResult();
+        return new JsonResult(StatusCode::SERVER_ERROR);
     }
 
     /**
@@ -88,10 +94,14 @@ class CollectionService
      * @param array $req
      * @return mixed
      */
-    public function getCollectionByUserSku(array $req)
+    public function checkCollectionByUserSku(array $req)
     {
-        $result = $this->collectionDao->findByUserSku($req["userId"], $req["skuId"]);
-        return $result;
+        if (empty($req["userId"]) || empty($req["skuId"])) return new JsonResult(StatusCode::PARAM_LACKED);
+        $collection = $this->collectionDao->findByUserSku($req["userId"], $req["skuId"]);
+        $result = new \stdClass();
+        if (empty($collection)) $result->isCollected = 0;
+        else $result->isCollected = 1;
+        return new JsonResult(StatusCode::SUCCESS, $result);
     }
 
     /**
